+++ src/components/comments/CommentsContainer.jsx
@@ -1,7 +1,7 @@
 import React, { Component } from 'react'
 import ShowComment from './ShowComment'
 import { animateScroll } from 'react-scroll'
-import "./scrollbar.css";
+import "../scrollbar.css";
 
 export class CommentsContainer extends Component {
     scrollToTop =() =>{
